+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## [2.0.3] dev-2020-10-07
 ### Added
+- New {filtercodes} tag. Note: Only works for teachers and above.
 - New {alert style}{/alert} tags (ALPHA).
 - New {ifincohort idname|idnumber}{/ifincohort} tags.
 - New {webpage} tag.